remove clutter + migrate code to dedicated files
